#1940 Renamed Task to Volunteer Task for Views and ViewModels
