+++ AllReadyApp/Web-App/AllReady.UnitTest/Areas/Admin/Controllers/TaskControllerTests.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using AllReady.Areas.Admin.Controllers;
 using AllReady.Areas.Admin.Features.Tasks;
-using AllReady.Areas.Admin.ViewModels.Task;
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
 using AllReady.Areas.Admin.ViewModels.Validators.Task;
 using AllReady.Extensions;
 using AllReady.UnitTest.Extensions;
