+++ AllReadyApp/Web-App/AllReady.UnitTest/Areas/Admin/Features/Tasks/CreateVolunteerTaskQueryHandlerShould.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using AllReady.Areas.Admin.Features.Tasks;
-using AllReady.Areas.Admin.ViewModels.Task;
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
 using AllReady.Models;
 using Xunit;
