+++ AllReadyApp/Web-App/AllReady.UnitTest/Areas/Admin/Features/Tasks/DetailsQueryHandlerShould.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using AllReady.Areas.Admin.Features.Tasks;
-using AllReady.Areas.Admin.ViewModels.Task;
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
 using AllReady.Models;
 using Xunit;
 using AllReady.Areas.Admin.ViewModels.Shared;
