+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Tasks/CreateVolunteerTaskQueryHandler.cs
@@ -1,6 +1,6 @@
 ﻿using System.Linq;
 using System.Threading.Tasks;
-using AllReady.Areas.Admin.ViewModels.Task;
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
 using AllReady.Models;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
