+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Tasks/DeleteQuery.cs
@@ -1,4 +1,4 @@
-﻿using AllReady.Areas.Admin.ViewModels.Task;
+﻿using AllReady.Areas.Admin.ViewModels.VolunteerTask;
 using MediatR;
 
 namespace AllReady.Areas.Admin.Features.Tasks
