+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Tasks/EditVolunteerTaskQueryHandler.cs
@@ -1,5 +1,5 @@
 ﻿using System.Threading.Tasks;
-using AllReady.Areas.Admin.ViewModels.Task;
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
 using AllReady.Models;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
