+++ AllReadyApp/Web-App/AllReady/Areas/Admin/ViewModels/Validators/Task/IValidateVolunteerTaskEditViewModelValidator.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using AllReady.Areas.Admin.ViewModels.Task;
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
 
 namespace AllReady.Areas.Admin.ViewModels.Validators.Task
 {
