+++ AllReadyApp/Web-App/AllReady/Areas/Admin/ViewModels/Validators/Task/VolunteerTaskEditViewModelValidator.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using AllReady.Areas.Admin.ViewModels.Task;
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
 using AllReady.Features.Events;
 using AllReady.Models;
 using MediatR;
